refactor: split renderer into focused modules with named constants
Extract main.js (399 lines) into four focused modules:
- lib/state.js: shared mutable state and logging
- lib/audio.js: AudioContext, RMS, vowel detection
- lib/visual.js: animation loop, blinking, lip-sync
- lib/comm.js: WebSocket, audio queue, mic streaming

main.js is now a thin orchestrator (~110 lines).
All magic numbers (FFT size, smoothing, RMS boost, vowel boundaries,
blink duration, silence floor) are extracted as named constants.
Also derives wsUrl from electron.port, removing the redundant setting.
Updated .gitignore to allow electron/renderer/lib/ despite global lib/ ignore.
