+++ .gitignore
@@ -15,6 +15,7 @@ downloads/
 eggs/
 .eggs/
 lib/
+!electron/renderer/lib/
 lib64/
 parts/
 sdist/
